+++ shell/src/main/java/com/samsung/command/testscript/TestScript2Command.java
@@ -1,7 +1,7 @@
 package com.samsung.command.testscript;
 
 import com.samsung.command.Command;
-import com.samsung.validator.ArgumentsValidator;
+import com.samsung.validator.CommandValidator;
 
 public class TestScript2Command implements Command {
     private final ScriptManager scriptManager;
@@ -12,7 +12,7 @@ public class TestScript2Command implements Command {
 
     @Override
     public void execute(String[] cmdArgs) {
-        ArgumentsValidator.validateOneArgs(cmdArgs);
+        CommandValidator.validateOneArgs(cmdArgs);
         boolean isTestPassed = scriptManager.testScript2();
         validatePostcondition(isTestPassed);
     }
