+++ shell/src/main/java/com/samsung/command/testshell/EraseCommand.java
@@ -2,7 +2,7 @@ package com.samsung.command.testshell;
 
 import com.samsung.command.ArgumentResolver;
 import com.samsung.command.Command;
-import com.samsung.validator.ArgumentsValidator;
+import com.samsung.validator.CommandValidator;
 
 public class EraseCommand implements Command {
 
@@ -17,7 +17,7 @@ public class EraseCommand implements Command {
 
     @Override
     public void execute(String[] cmdArgs) {
-        ArgumentsValidator.validateThreeArgs(cmdArgs);
+        CommandValidator.validateThreeArgs(cmdArgs);
         testShellManager.erase(
                 extractValidatedEraseLba(cmdArgs),
                 extractValidatedEraseSize(cmdArgs));
