+++ shell/src/main/java/com/samsung/command/testshell/FullWriteCommand.java
@@ -2,7 +2,7 @@ package com.samsung.command.testshell;
 
 import com.samsung.command.ArgumentResolver;
 import com.samsung.command.Command;
-import com.samsung.validator.ArgumentsValidator;
+import com.samsung.validator.CommandValidator;
 
 public class FullWriteCommand implements Command {
 
@@ -16,7 +16,7 @@ public class FullWriteCommand implements Command {
 
     @Override
     public void execute(String[] cmdArgs) {
-        ArgumentsValidator.validateTwoArgs(cmdArgs);
+        CommandValidator.validateTwoArgs(cmdArgs);
         testShellManager.fullwrite(extractValidatedValue(cmdArgs));
     }
