+++ shell/src/main/java/com/samsung/command/testshell/ReadCommand.java
@@ -2,7 +2,7 @@ package com.samsung.command.testshell;
 
 import com.samsung.command.ArgumentResolver;
 import com.samsung.command.Command;
-import com.samsung.validator.ArgumentsValidator;
+import com.samsung.validator.CommandValidator;
 
 public class ReadCommand implements Command {
 
@@ -16,7 +16,7 @@ public class ReadCommand implements Command {
 
     @Override
     public void execute(String[] cmdArgs) {
-        ArgumentsValidator.validateTwoArgs(cmdArgs);
+        CommandValidator.validateTwoArgs(cmdArgs);
         testShellManager.read(extractValidatedLba(cmdArgs));
     }
